requirements.txt
boto3==1.38.8
numpy==2.2.5
polars==1.29.0
python-dotenv==1.1.0
python-telegram-bot==22.0
pytz==2025.2
torch==2.2.0
transformers==4.51.3
